refactor(saga): Phase 2b move domain to Domain/{OrderSaga,RefundSaga,}/
Pure relocation + namespace rename. State machines remain static and
pure. Consumer usings updated; Migrations entity names rebound to new
namespaces to keep ModelSnapshot drift-free.
