+++ saga-microservice/Saga.Service/Domain/OrderSaga/OrderSagaState.cs
@@ -1,4 +1,4 @@
-namespace Saga.Service.Models;
+namespace Saga.Service.Domain.OrderSaga;
 
 internal sealed class OrderSagaState
 {
