+++ saga-microservice/Saga.Service/Domain/RefundSaga/RefundSagaState.cs
@@ -1,4 +1,4 @@
-namespace Saga.Service.Models;
+namespace Saga.Service.Domain.RefundSaga;
 
 internal sealed class RefundSagaState
 {
