+++ saga-microservice/Saga.Service/Domain/RefundSaga/RefundSagaStep.cs
@@ -1,4 +1,4 @@
-namespace Saga.Service.Models;
+namespace Saga.Service.Domain.RefundSaga;
 
 internal enum RefundSagaStep
 {
