+++ saga-microservice/Saga.Service/Domain/SagaStatus.cs
@@ -1,4 +1,4 @@
-namespace Saga.Service.Models;
+namespace Saga.Service.Domain;
 
 internal enum SagaStatus
 {
