+++ saga-microservice/Saga.Service/IntegrationEvents/EventHandlers/OrderSagaReplyProcessor.cs
@@ -2,11 +2,11 @@ using ECommerce.Shared.Infrastructure.EventBus;
 using ECommerce.Shared.Infrastructure.Outbox;
 using Microsoft.EntityFrameworkCore;
 using Saga.Service.Contracts.Integration.InboundEvents;
+using Saga.Service.Domain;
+using Saga.Service.Domain.OrderSaga;
 using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.Infrastructure.Reaper;
-using Saga.Service.Models;
 using Saga.Service.Observability;
-using Saga.Service.StateMachines;
 
 namespace Saga.Service.IntegrationEvents.EventHandlers;
