+++ saga-microservice/Saga.Tests/Api/SagaObservabilityTests.cs
@@ -6,9 +6,9 @@ using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Saga.Service.Contracts.Integration.InboundEvents;
+using Saga.Service.Domain.OrderSaga;
 using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.IntegrationEvents.EventHandlers;
-using Saga.Service.Models;
 using Saga.Service.Observability;
 
 namespace Saga.Tests.Api;
