+++ saga-microservice/Saga.Tests/Domain/RefundSagaStateMachineTests.cs
@@ -1,7 +1,7 @@
 using ECommerce.Shared.IntegrationEvents.Commands;
 using Saga.Service.Contracts.Integration.InboundEvents;
-using Saga.Service.Models;
-using Saga.Service.StateMachines;
+using Saga.Service.Domain;
+using Saga.Service.Domain.RefundSaga;
 
 namespace Saga.Tests.Domain;
