+++ saga-microservice/Saga.Tests/EndToEnd/OrderSagaEndToEndTests.cs
@@ -5,9 +5,10 @@ using ECommerce.Shared.IntegrationEvents.Commands;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Saga.Service.Contracts.Integration.InboundEvents;
+using Saga.Service.Domain;
+using Saga.Service.Domain.OrderSaga;
 using Saga.Service.Infrastructure.Data.EntityFramework;
 using Saga.Service.IntegrationEvents.EventHandlers;
-using Saga.Service.Models;
 
 namespace Saga.Tests.EndToEnd;
